lastfix

+++ Isaac-AnomalyService/Models/WeatherApiWield.cs
@@ -8,6 +8,6 @@ namespace Isaac_AnomalyService.Models
     public class WeatherApiWield
     {
         public double ApiTemp { get; set; }
-        public int ApiHum { get; set; }
+        public double ApiHum { get; set; }
     }
 }
